+++ layer_styles/DEM Netherlands.qml
@@ -1,0 +1,27 @@
+<!DOCTYPE qgis PUBLIC 'http://mrcc.com/qgis.dtd' 'SYSTEM'>
+<qgis version="2.14.20-Essen" minimumScale="0" maximumScale="1e+08" hasScaleBasedVisibilityFlag="0">
+  <pipe>
+    <rasterrenderer opacity="1" alphaBand="-1" classificationMax="57.8495" classificationMinMaxOrigin="CumulativeCutFullExtentEstimated" band="1" classificationMin="-4.4585" type="singlebandpseudocolor">
+      <rasterTransparency/>
+      <rastershader>
+        <colorrampshader colorRampType="INTERPOLATED" clip="0">
+          <item alpha="255" value="-5.723" label="-5.723000" color="#218394"/>
+          <item alpha="255" value="-1.76" label="-1.760000" color="#4d9b9d"/>
+          <item alpha="255" value="-0.608" label="-0.608000" color="#79b4a7"/>
+          <item alpha="255" value="0.245" label="0.245000" color="#a5ccb1"/>
+          <item alpha="255" value="1.097" label="1.097000" color="#d1e5bb"/>
+          <item alpha="255" value="3.029" label="3.029000" color="#fefec5"/>
+          <item alpha="255" value="6.753" label="6.753000" color="#eadca2"/>
+          <item alpha="255" value="10.694" label="10.694000" color="#d6ba7f"/>
+          <item alpha="255" value="16.294" label="16.294000" color="#c3985d"/>
+          <item alpha="255" value="26.858" label="26.858000" color="#af763a"/>
+          <item alpha="255" value="243.48" label="243.480000" color="#9c5518"/>
+        </colorrampshader>
+      </rastershader>
+    </rasterrenderer>
+    <brightnesscontrast brightness="0" contrast="0"/>
+    <huesaturation colorizeGreen="128" colorizeOn="0" colorizeRed="255" colorizeBlue="128" grayscaleMode="0" saturation="0" colorizeStrength="100"/>
+    <rasterresampler maxOversampling="2"/>
+  </pipe>
+  <blendMode>0</blendMode>
+</qgis>