+++ layer_styles/Land use.qml
@@ -1,0 +1,86 @@
+<!DOCTYPE qgis PUBLIC 'http://mrcc.com/qgis.dtd' 'SYSTEM'>
+<qgis version="2.14.20-Essen" minimumScale="0" maximumScale="1e+08" hasScaleBasedVisibilityFlag="0">
+  <pipe>
+    <rasterrenderer opacity="1" alphaBand="-1" classificationMax="193" classificationMinMaxOrigin="CumulativeCutFullExtentEstimated" band="1" classificationMin="26" type="singlebandpseudocolor">
+      <rasterTransparency/>
+      <rastershader>
+        <colorrampshader colorRampType="DISCRETE" clip="0">
+          <item alpha="255" value="2" label="&quot;BAG - woonfunctie&quot;" color="#ac0000"/>
+          <item alpha="255" value="3" label="&quot;BAG - celfunctie&quot;" color="#4d5193"/>
+          <item alpha="255" value="4" label="&quot;BAG - industriefunctie&quot;" color="#c787ff"/>
+          <item alpha="255" value="5" label="&quot;BAG - kantoorfunctie&quot;" color="#9c5fd0"/>
+          <item alpha="255" value="6" label="&quot;BAG - winkelfunctie&quot;" color="#ffc9b0"/>
+          <item alpha="255" value="7" label="&quot;BAG - kassen&quot;" color="#e09cf1"/>
+          <item alpha="255" value="8" label="&quot;BAG - logiesfunctie&quot;" color="#9d8cff"/>
+          <item alpha="255" value="9" label="&quot;BAG - bijeenkomstfunctie&quot;" color="#674093"/>
+          <item alpha="255" value="10" label="&quot;BAG - bag sportfunctie&quot;" color="#f57dff"/>
+          <item alpha="255" value="11" label="&quot;BAG - onderwijsfunctie&quot;" color="#c8ede6"/>
+          <item alpha="255" value="12" label="&quot;BAG - gezondheidsfunctie&quot;" color="#a4d8cd"/>
+          <item alpha="255" value="13" label="&quot;BAG - overig kleiner dan 50 m2" color="#7aa2a8"/>
+          <item alpha="255" value="14" label="&quot;BAG - overig groter dan 50 m2&quot;" color="#6a8991"/>
+          <item alpha="255" value="16" label="&quot;CBS - woongebied&quot;" color="#d94747"/>
+          <item alpha="255" value="17" label="&quot;CBS - winkelgebied&quot;" color="#ffcd69"/>
+          <item alpha="255" value="18" label="&quot;CBS - openbaar terrein&quot;" color="#ffcd69"/>
+          <item alpha="255" value="19" label="&quot;CBS - bedrijventerrein&quot;" color="#aaaaff"/>
+          <item alpha="255" value="20" label="&quot;OSM - sportcentrum&quot;" color="#df72e9"/>
+          <item alpha="255" value="21" label="&quot;OSM - stadion&quot;" color="#c162c9"/>
+          <item alpha="255" value="22" label="&quot;OSM - sportveld&quot;" color="#3e9661"/>
+          <item alpha="255" value="23" label="&quot;OSM - atletiekbaan&quot;" color="#8f6600"/>
+          <item alpha="255" value="24" label="&quot;OSM - tennisbaan&quot;" color="#a755ae"/>
+          <item alpha="255" value="25" label="&quot;OSM - sportbaan&quot;" color="#944c9b"/>
+          <item alpha="255" value="27" label="&quot;BRP/Top10 - gras&quot;" color="#b9ff98"/>
+          <item alpha="255" value="34" label="&quot;Top10 - spoor&quot;" color="#000000"/>
+          <item alpha="255" value="39" label="&quot;Top10 - bos / natuur&quot;" color="#008b00"/>
+          <item alpha="255" value="44" label="&quot;BRP - Aardappelen&quot;" color="#cb8e00"/>
+          <item alpha="255" value="50" label="&quot;BRP - Bieten&quot;" color="#ff69ff"/>
+          <item alpha="255" value="52" label="&quot;BRP - Granen&quot;" color="#ffd500"/>
+          <item alpha="255" value="53" label="&quot;BRP - Bloembollen&quot;" color="#e4be00"/>
+          <item alpha="255" value="54" label="&quot;BRP - Bloemen&quot;" color="#ff0000"/>
+          <item alpha="255" value="55" label="&quot;BRP - Peulvruchten&quot;" color="#bb9c00"/>
+          <item alpha="255" value="59" label="&quot;BRP - Kwekerij&quot;" color="#9b8100"/>
+          <item alpha="255" value="60" label="&quot;BRP/CBS - bos&quot;" color="#0a6400"/>
+          <item alpha="255" value="65" label="&quot;BRP - Braakliggend&quot;" color="#755f00"/>
+          <item alpha="255" value="66" label="&quot;BRP - Knol en Bolgewassen&quot;" color="#ffb700"/>
+          <item alpha="255" value="67" label="&quot;BRP - Overige akkerbouwgewassen&quot;" color="#de9f00"/>
+          <item alpha="255" value="70" label="&quot;BRP - Natuur&quot;" color="#8abc71"/>
+          <item alpha="255" value="71" label="&quot;BRP - Fruitteelt&quot;" color="#42d64c"/>
+          <item alpha="255" value="72" label="&quot;BRP - Hoogstam&quot;" color="#694b00"/>
+          <item alpha="255" value="81" label="&quot;BRP - Groenbemesters&quot;" color="#34ffa4"/>
+          <item alpha="255" value="94" label="&quot;BRP - Maïs&quot;" color="#c88f00"/>
+          <item alpha="255" value="118" label="&quot;CBS/Top10 - begraafplaats&quot;" color="#580058"/>
+          <item alpha="255" value="119" label="&quot;CBS - glastuinbouwterrein&quot;" color="#f5c0ff"/>
+          <item alpha="255" value="120" label="&quot;CBS - erf&quot;" color="#777fea"/>
+          <item alpha="255" value="121" label="&quot;CBS - park en plantsoen&quot;" color="#3bc044"/>
+          <item alpha="255" value="122" label="&quot;CBS - volkstuin&quot;" color="#00cfad"/>
+          <item alpha="255" value="123" label="&quot;CBS - dagrecreatief terrein&quot;" color="#a7ffdb"/>
+          <item alpha="255" value="124" label="&quot;CBS - verblijfsrecreatief terrein&quot;" color="#89dbc4"/>
+          <item alpha="255" value="125" label="&quot;CBS - stortplaats&quot;" color="#82c9bc"/>
+          <item alpha="255" value="126" label="&quot;CBS - wrakkenopslagplaats&quot;" color="#6caa8e"/>
+          <item alpha="255" value="128" label="&quot;CBS - delfstofwinplaats&quot;" color="#cfefff"/>
+          <item alpha="255" value="129" label="&quot;CBS - bouwterrein&quot;" color="#1a4c51"/>
+          <item alpha="255" value="130" label="&quot;CBS - semi-verhard overig terrein&quot;" color="#e6c9e3"/>
+          <item alpha="255" value="131" label="&quot;Top10 - transformatorstation&quot;" color="#ff6e70"/>
+          <item alpha="255" value="132" label="&quot;Top10 - watertank&quot;" color="#00b3ff"/>
+          <item alpha="255" value="133" label="&quot;Top10 - kas&quot;" color="#6beac6"/>
+          <item alpha="255" value="134" label="&quot;CBS - open droog natuurlijk terrein&quot;" color="#55aa7f"/>
+          <item alpha="255" value="135" label="&quot;CBS - open nat natuurlijk terrein&quot;" color="#008684"/>
+          <item alpha="255" value="144" label="&quot;CBS/Top10 - binnenwater&quot;" color="#0000ff"/>
+          <item alpha="255" value="146" label="&quot;Top10 - berm&quot;" color="#02944b"/>
+          <item alpha="255" value="147" label="&quot;CBS - spoorberm&quot;" color="#9f3b01"/>
+          <item alpha="255" value="148" label="&quot;CBS - wegberm&quot;" color="#00aa71"/>
+          <item alpha="255" value="149" label="&quot;CBS - vliegveld&quot;" color="#b78586"/>
+          <item alpha="255" value="150" label="&quot;CBS - sportterrein&quot;" color="#ff7dcb"/>
+          <item alpha="255" value="249" label="&quot;OSM - parkeerterrein&quot;" color="#e1aeaf"/>
+          <item alpha="255" value="251" label="&quot;Top10 - primaire weg&quot;" color="#404040"/>
+          <item alpha="255" value="252" label="&quot;Top10 - secundaire weg&quot;" color="#828282"/>
+          <item alpha="255" value="253" label="&quot;Top10 - tertiaire weg&quot;" color="#c5c5c5"/>
+          <item alpha="255" value="254" label="&quot;Top10 - overige weg&quot;" color="#e1e1e1"/>
+        </colorrampshader>
+      </rastershader>
+    </rasterrenderer>
+    <brightnesscontrast brightness="0" contrast="0"/>
+    <huesaturation colorizeGreen="128" colorizeOn="0" colorizeRed="255" colorizeBlue="128" grayscaleMode="0" saturation="0" colorizeStrength="100"/>
+    <rasterresampler maxOversampling="2"/>
+  </pipe>
+  <blendMode>0</blendMode>
+</qgis>