+++ layer_styles/Rain 5min.qml
@@ -1,0 +1,28 @@
+<!DOCTYPE qgis PUBLIC 'http://mrcc.com/qgis.dtd' 'SYSTEM'>
+<qgis version="2.14.20-Essen" minimumScale="-4.65661e-10" maximumScale="1e+08" hasScaleBasedVisibilityFlag="0">
+  <pipe>
+    <rasterrenderer opacity="1" alphaBand="0" classificationMax="0.0198" classificationMinMaxOrigin="CumulativeCutFullExtentEstimated" band="1" classificationMin="0" type="singlebandpseudocolor">
+      <rasterTransparency/>
+      <rastershader>
+        <colorrampshader colorRampType="INTERPOLATED" clip="0">
+          <item alpha="0" value="0" label="0.000000" color="#ffffff"/>
+          <item alpha="255" value="0.0009" label="0.000900" color="#ffffff"/>
+          <item alpha="255" value="0.002171" label="0.002171" color="#e9ebec"/>
+          <item alpha="255" value="0.005236" label="0.005236" color="#cfd3d6"/>
+          <item alpha="255" value="0.01263" label="0.012630" color="#aabdc9"/>
+          <item alpha="255" value="0.030465" label="0.030465" color="#5ea5d4"/>
+          <item alpha="255" value="0.073485" label="0.073485" color="#3091d1"/>
+          <item alpha="255" value="0.177251" label="0.177251" color="#2982bb"/>
+          <item alpha="255" value="0.427542" label="0.427542" color="#7d9c70"/>
+          <item alpha="255" value="1.031263" label="1.031263" color="#f0ba12"/>
+          <item alpha="255" value="2.487484" label="2.487484" color="#c0527a"/>
+          <item alpha="255" value="6" label="6.000000" color="#9b59b6"/>
+        </colorrampshader>
+      </rastershader>
+    </rasterrenderer>
+    <brightnesscontrast brightness="0" contrast="0"/>
+    <huesaturation colorizeGreen="128" colorizeOn="0" colorizeRed="255" colorizeBlue="128" grayscaleMode="0" saturation="0" colorizeStrength="100"/>
+    <rasterresampler maxOversampling="2"/>
+  </pipe>
+  <blendMode>0</blendMode>
+</qgis>