+++ layer_styles/Rain day.qml
@@ -1,0 +1,28 @@
+<!DOCTYPE qgis PUBLIC 'http://mrcc.com/qgis.dtd' 'SYSTEM'>
+<qgis version="2.14.20-Essen" minimumScale="0" maximumScale="1e+08" hasScaleBasedVisibilityFlag="0">
+  <pipe>
+    <rasterrenderer opacity="1" alphaBand="-1" classificationMax="0.0198" classificationMinMaxOrigin="CumulativeCutFullExtentEstimated" band="1" classificationMin="0" type="singlebandpseudocolor">
+      <rasterTransparency/>
+      <rastershader>
+        <colorrampshader colorRampType="INTERPOLATED" clip="0">
+          <item alpha="0" value="0" label="0.000000" color="#ffffff"/>
+          <item alpha="255" value="0.049" label="0.049000" color="#ffffff"/>
+          <item alpha="255" value="0.09816" label="0.098160" color="#ecedee"/>
+          <item alpha="255" value="0.19664" label="0.196640" color="#d7dbdd"/>
+          <item alpha="255" value="0.39392" label="0.393920" color="#c4c9cd"/>
+          <item alpha="255" value="0.789125" label="0.789125" color="#8fb4cd"/>
+          <item alpha="255" value="1.580823" label="1.580823" color="#499ed7"/>
+          <item alpha="255" value="3.166799" label="3.166799" color="#4b8b9c"/>
+          <item alpha="255" value="6.343924" label="6.343924" color="#b1ae45"/>
+          <item alpha="255" value="12.708531" label="12.708531" color="#ed9a1e"/>
+          <item alpha="255" value="25.458497" label="25.458497" color="#cf4f62"/>
+          <item alpha="255" value="51" label="51.000000" color="#9b59b6"/>
+        </colorrampshader>
+      </rastershader>
+    </rasterrenderer>
+    <brightnesscontrast brightness="0" contrast="0"/>
+    <huesaturation colorizeGreen="128" colorizeOn="0" colorizeRed="255" colorizeBlue="128" grayscaleMode="0" saturation="0" colorizeStrength="100"/>
+    <rasterresampler maxOversampling="2"/>
+  </pipe>
+  <blendMode>0</blendMode>
+</qgis>