+++ layer_styles/Rain hour.qml
@@ -1,0 +1,28 @@
+<!DOCTYPE qgis PUBLIC 'http://mrcc.com/qgis.dtd' 'SYSTEM'>
+<qgis version="2.14.20-Essen" minimumScale="-4.65661e-10" maximumScale="1e+08" hasScaleBasedVisibilityFlag="0">
+  <pipe>
+    <rasterrenderer opacity="1" alphaBand="0" classificationMax="193" classificationMinMaxOrigin="CumulativeCutFullExtentEstimated" band="1" classificationMin="26" type="singlebandpseudocolor">
+      <rasterTransparency/>
+      <rastershader>
+        <colorrampshader colorRampType="INTERPOLATED" clip="0">
+          <item alpha="0" value="0" label="0.000000" color="#ffffff"/>
+          <item alpha="255" value="0.019" label="0.019000" color="#ffffff"/>
+          <item alpha="255" value="0.039812" label="0.039812" color="#ebedee"/>
+          <item alpha="255" value="0.083421" label="0.083421" color="#d6d9dc"/>
+          <item alpha="255" value="0.174798" label="0.174798" color="#d6d9dc"/>
+          <item alpha="255" value="0.366266" label="0.366266" color="#62a6d4"/>
+          <item alpha="255" value="0.767463" label="0.767463" color="#2c88c5"/>
+          <item alpha="255" value="1.60812" label="1.608120" color="#5c918c"/>
+          <item alpha="255" value="3.369607" label="3.369607" color="#bab13d"/>
+          <item alpha="255" value="7.060575" label="7.060575" color="#e85d35"/>
+          <item alpha="255" value="14.79452" label="14.794520" color="#c05279"/>
+          <item alpha="255" value="31" label="31.000000" color="#9b59b6"/>
+        </colorrampshader>
+      </rastershader>
+    </rasterrenderer>
+    <brightnesscontrast brightness="0" contrast="0"/>
+    <huesaturation colorizeGreen="128" colorizeOn="0" colorizeRed="255" colorizeBlue="128" grayscaleMode="0" saturation="0" colorizeStrength="100"/>
+    <rasterresampler maxOversampling="2"/>
+  </pipe>
+  <blendMode>0</blendMode>
+</qgis>